+++ tools/efrotools/toolconfig.py
@@ -183,6 +183,7 @@ def _filter_tool_config(projroot: Path, cfg: str) -> str:
     strict_equality = True
     local_partial_types = True
     no_implicit_reexport = True
+    fixed_format_cache = True
 
     enable_error_code = redundant-expr, truthy-bool, \
 truthy-function, unused-awaitable, explicit-override